Subida diaria a develop

+++ Backend/Server/Server/Models/Game.cs
@@ -3,7 +3,7 @@
     public class Game
     {
         public int Id { get; set; }
-        public int Time { get; set; }
+        public double Time { get; set; }
         public ICollection<GameInfo> gameInfos { get; set; } = new List<GameInfo>();
     }
 }
